% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lond}
\alias{lond}
\title{Grid coordinates for Upper Tier Local Authorities in London}
\format{
A data frame with 33 rows and 4 variables:
\describe{
  \item{code}{ONS geography code}
  \item{name}{ONS geography name}
  \item{col}{integer grid reference for where to place local authority on x axis}
  \item{row}{integer grid reference for where to place local authority on y axis}
}
}
\usage{
lond
}
\description{
A dataset containing x and y grid references for local authorities in London
}
\keyword{datasets}
