% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{london_shp}
\alias{london_shp}
\title{A shapefile of low resolution boundary for London (2021) from the ONS geography portal}
\format{
an sf object with 1 row and 12 variables
\describe{
  \item{RGN21CD}{ONS geography code}
  \item{RGN21NM}{Region name}
  \item{BNG_E}{Eastings}
  \item{BNG_N}{Northings}
  \item{LONG}{Longitude}
  \item{LAT}{Latitude}
  \item{OBJECTID}{Object ID}
  \item{Shape__Area}{Area, m2}
  \item{Shape__Length}{m}
  \item{GlobalID}{Global ID}
  \item{geometry}{geometry}
}
}
\usage{
london_shp
}
\description{
A shapefile of low resolution boundary for London (2021) from the ONS geography portal
}
\keyword{datasets}
