% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outside_london_grid}
\alias{outside_london_grid}
\title{Grid coordinates for Upper Tier Local Authorities outside of London}
\format{
A data frame with 122 rows and 4 variables:
\describe{
  \item{code}{ONS geography code}
  \item{name}{ONS geography name}
  \item{col}{integer grid reference for where to place local authority on x axis}
  \item{row}{integer grid reference for where to place local authority on y axis}
}
}
\usage{
outside_london_grid
}
\description{
A dataset containing x and y grid references for local authorities outside of
London
}
\keyword{datasets}
