% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{significance_colours}
\alias{significance_colours}
\title{A colour scheme for charts with labels for "Better", "Same", "Worse", and
"Not compared"}
\format{
a named vector with 4 items
}
\usage{
significance_colours
}
\description{
A colour scheme based on the RAG colours in Fingertips
}
\keyword{datasets}
