% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{utla_plot}
\alias{utla_plot}
\title{Map of charts faceted by Upper Tier Local Authority in England}
\usage{
utla_plot(
  london_data,
  england_data,
  london_grid,
  england_grid,
  output_filename
)
}
\description{
Map of charts faceted by Upper Tier Local Authority in England
}
\details{
includes an inset for London
}
